% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tLagPropOdds.R
\name{print}
\alias{print}
\alias{print.tLagObj}
\title{Print Analysis Results}
\usage{
\method{print}{tLagObj}(x, ...)
}
\arguments{
\item{x}{A tLagObj object. The value returned by tLagPropOdds().}

\item{...}{Ignored.}
}
\description{
Prints the key results.
}
\examples{

data(tLagData)

# full AIPWCC estimator
res <- tLagPropOdds(data = tLagData, 
                    ti = "x",  
                    td = c("hospStatus", "daysOut"))

print(x = res)
}
